generate: split the generate code
In preparation for multiple commits. The fmt subcommand will reuse the
package loading code, so leave that in loader.

For #37.
